Add computed column `DisplayName` to `Author` entity
Introduced a computed column `DisplayName` to the `Author` entity in `AppDbContext`, which concatenates `FirstName` and `LastName` into a read-only column. Updated the `Author` class to include the `DisplayName` property, marked as required with a max length of 100 characters.

Created migration `20251202142933_AddDisplayName` to add the `DisplayName` column to the `Authors` table, with the SQL formula `[FirstName] + ' ' + [LastName]`. Updated `AppDbContextModelSnapshot` and generated the corresponding designer file to reflect the schema changes.

Demonstrated the addition of a new `Author` instance in `Program.cs` to validate the changes.